Wed, Mar 9, 2022, 8:24 PM +00:00

+++ finance/app.py
@@ -86,7 +86,7 @@ def buy():
         shares = request.form.get("shares")
         user_id = session["user_id"]
         cash = int(db.execute("SELECT cash FROM users WHERE id = ?", user_id)[0]["cash"])
-        new_cash = int(cash - price * shares)
+        new_cash = cash - price * shares
 
         if new_cash < 0:
             return apology("You gonna need more cash for that")
